drivers: mailbox: add DT parameter to determine the callback context
The mailboxes client callback can be called under interrupt context or in
normal context. The need depends on the client.
For instance a virtio I2C client needs to be called under interrupt
context. Indeed the associated bus controller should allows an I2C drivers
to access to the hardware during the probe ( atomic context).
In such case the remote device "Ack" reception have to be to be treated
under interrupt context.
On the other hand a virtio rpmsg should not work in interrupt context. It
mix several services in one virtio.

This patch add the capability to call the mailbox client under interrupt
or normal context depending on the mailbox channel ID.
- if mailbox channel ID is from 0 to 0xFF the client is called in normal
  context
- if mailbox channel ID is from 0x100 to 0x1FF the client is called
  in interrupt context

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: I7baa246a050c8204ee61add492641791e984fd17