dma-direct: prevent incorrect check of min dma address
When mapping dma addresses, this check will try to convert the address :
0xC0000000 using phys_to_dma,
thinking it is the smallest dma address possible.

But this call return 0xFFFFFFFF on an stm32mp1,
since it's not a dma address at all,
this results in rejecting any dma address.

This patch is a temporary solution to prevent 0xFFFFFFFF,
from being considered the smallest address possible.

Signed-off-by: Nicolas Granger <nicolas.granger01@st.com>